Serious Edits
I made some serious changes to the UI as well as the back end of the
app. Events and Cameras can now be stored in an SQLi database, where
they are taken out and listed in the Add Event and Add Cameras pages. I
apologize if the code is confusing, if there is something you want me to
re-organize, I can do that.